refactor: Move handleDelete to PostControl component
