+++ src/components/PostControl.tsx
@@ -1,18 +1,31 @@
-import { Link } from "react-router-dom";
+import { Link, useNavigate } from "react-router-dom";
 import { FormEvent } from "react";
 import Button from "./Button";
 import { IPost } from "../@types/posts";
+import { deletePost } from "../api/posts";
+import toast from "../utils/toast";
 import "./PostControl.css";
 
-export function PostControl({
-  id,
-  post,
-  handleDelete,
-}: {
-  id: string;
-  post: IPost;
-  handleDelete: (event: FormEvent) => void;
-}) {
+export function PostControl({ id, post }: { id: string; post: IPost }) {
+  const go = useNavigate();
+
+  const handleDelete = async (event: FormEvent) => {
+    event.preventDefault();
+
+    if (!confirm("정말 삭제하시겠습니까?")) {
+      return;
+    }
+
+    const { response, data } = await deletePost(id);
+
+    if (!response.ok) {
+      toast(data.message);
+      return;
+    }
+
+    go("/");
+  };
+
   return (
     <div className="post__control">
       <Link to={`/posts/update/${id}`} state={post}>
